Copy from other repo
This Python code will solve the Fair and Square problem in CodeJam 2013's Qualification Round C.

The official information for this problem can be found at: https://code.google.com/codejam/contest/2270488/dashboard#s=p2

This code isn't optimal enough to handle the large sets, but it will take care of the small set. I've started work on a method to increase the efficiency, though it also involves pregenerating the possible Fair and Square values before running based on the limits on the problem's page. I'm not sure if that is in spirit of the problem. On to the description for this code.

I take in the range values and convert them to integers (which I think causes problems in the larger ranges associated with with larger data sets.

I then run through each value in the range and if it is a perfect square (as checked to see if the square of the int of it's squareroot equals itself and if it is a plaindrome (by use of a premade function that converts to a string and compares the s to s[::-1], the inverted string) it will append the value to a list.

Then it will do the exact same thing to the new list, created a shorter list. The length of this shorter list is then properly formatted and outputted to the dataout file.

Some thoughts on optimization to try:

The max range is given on the problem page. I also know that the Fair and Square values are constant so I can generate a list of valid ones at the start of the program. This could work for 1014, but is too time consuming to do at 10100.

Another thing to note is that I only need to work on values that are less than the fourth root of the max value. This is because all items in the list must by perfect squares of perfect squares. So when generating, I can work with this much much smaller range.

I found a list of all perfect square palindromes up to 55 digits at: http://fengyuan.com/palindrome.html . I can use this to check if the square root is a perfect square and a palindrome. The list of useable values is around 262.

In Python 2.7, I need to use xrange() to dynamically generate the values as the list is too long.

XXXXXXXXXXXXXXXXX I have to take into account the representations of long integers as they throw off the palindrome function. I think this occurs around 4008004 as that seems to be the highest value created with my code. The code now takes into account Long Integer format of adding an L to the end.

XXXXXXXXXXXXXXXXX I may be able to save time by only checking if the first half of the vlue matches the reverse of the second half. This should half the checks in that section. I have confirmed that this is always slower to do for values up to 10**7. The rate is about constant so I assume it is always slower.