Copy from other repo
This Python code will solve the Tic-Tac-Toe-Tomek problem in CodeJam 2013's Qualification Round A.

For more on the challenge: https://code.google.com/codejam/contest/2270488/dashboard#s=p0

To solve this problem, I defined four functions that could handle checking each given round.

The first function (check_matrixH) checks whether any of the four rows have a win condition in them as defined in the global h_answersX and h_answersO variables. If any row exists in the win conditions, the function returns "X" or "O" depending on the which set the row is found in. If it fails to find a win condition, the function returns '', a blank string.

The second function (check_matrixV) is almost the same as check_matrixH, except that it takes the transpose of the round. It will also join each row into a single string. Once again, it checks each row to see if it is a win condition and returns the appropriate result.

The third function (check_matrixD) takes in the four rows and creates two strings created from the diagonal values in the rows. And like before, it will check if they are win conditions and return the appropriate value ('X', 'O', '').

The fourth function (check_matrixT) also takes in the four rows and condenses them into a single string. It will then check whether there are any blank places ('.'). If blanks are found, it returns a 'G' representing that the game is still in progress. If no '.' are found, it returns a 'D' to indicate that the game was played to a draw.

A clever approach to this was how to efficiently run the four functions. Clearly, check_matrixT needed to be ran last as it is prone to false results as a win can contain '.'. While the first three functions can be used in any order, I choose to go with the order I wrote them. I created a results variable and left it as a blank string. I then ran each function, appended it's result to the results variable, and then checked to see if the length increased. By returning '' when a win condition is not found, the length of the results variable would not increase. Only when the three functions that can find a win condition have failed to do so will the fourth function run and decide if the game was tied or unfinished.

Once the results of the round are discovered, the program will then output a properly configured line to the output file. It will run through each test case and properly do this in a decently quick manner. The code successfully computed the answers to both the small and large datasets.