Copy from other repo
This Python code will solve the Bullseye challange in CodeJam 2013's Round 1A A. Well, it solves the Small set, the Large set takes too long.

The problem is located at: https://code.google.com/codejam/contest/2418487/dashboard#s=p0

The logic behind this solution is that you work your way down from the given tVal. To do this, I creted a while loop that runs as long as tVal remains positive and then I make sure to subtract from it each cycle.

The value that is subtracted is the area of the next ring. Starting from the provided rVal, the ring must be of value Pi*((rVal+1)^2 - rval^2). However, since the paint used is already measured in Pi cm^2, I can factor out Pi in my equation and not have to import Math or assign my own value to Pi. This reduces the math down to ((rVal+1)^2 - rval^2).

Each cycle, the value of tVal is reduced by the next ringsize, the value of tVal is checked and if still greater than 0 a count (of rings so far) is incremented by one. The last step is to increment rVal by 2. This is needed to account for the painted ring and the blank ring that we skip.

A downside in using Python is that it simply cannot perform fast enough for some of the larger values provided (such as starting at a rVal of 1 and using up tVal of 100000000). Python's inability to handle these large operations presents a severe handicap while undertaking the CodeJam challenges.

A thought on optimization is to create a list of all possible ring size values and then iterate over that list while checking the test cases. I attempted to make a list of values but it was going to take a couple hours to generate in Python.