Copy from other repo
This Python code will solve the Good Luck challenge in CodeJam 2013's Round 1A C.

This will solve the C-Small-Practice-1.in at https://code.google.com/codejam/contest/2418487/dashboard#s=p2 . Given enough time, it will also solve the larger problem set. Python seems too slow to handle the harder challenges.

The challenge involves correctly guessing the correct numbers used to create the products you are given as clues.

To simplify the coding, this program uses some methods to organize the work.

The first method, called makeGuessList, creates a list of every possible combination of numbers that could be used. It takes in the number of values (length) and the maximum value for a digit. We know how many numbers are used, and we know that the range is between 2 and some number given. I hardcoded a series of for-loops to create this list.

The next step is a method called makeProductList. It takes in the provided guessList from the previous method and the number of values (length). It then generate a list of all possible products for that guess. This is accomplished using my binary list comprehension. Then going through each member of guessList, part by part it generates every possible combination of products and appends them to a list in the same order. As both the binary list comprehension and the guessList are strings, I can easily work digit by digit for each and on the fly find the product of each, and then sum that.

With the brunt of the work completed in advance, I am able to go through and use my findAnswer method to the best guess for what the original numbers are. Taking in the created productList, guessList, and the unique herProducts provided for each scenario, I can do some calculations to find the best guess. I begin by adding a bunch of 1's to the end of each item in productList (just in case I get a bunch of 1s in the provided hints). I then run through each item in the productList and check if any of the provided herProducts are found. If they are found, I remove that item from the sublist in herProducts and if not I heavily penalize the guess by adding a bunch of zeroes. Next, a create a list of the length for each sublist in the modified productList. The smallest value will have had the least additions of 0s and the most items removed, so I find the index of that smallest value and look up what guess resides there in guessList.

This solution had some problems during implementation where I tried to create a fresh copy of productList using the old newList = oldList[:] trick, but for some reason it would not update and I would get a lot of errors and null entries, so now I just generate the productList for each round of guessing. This is much less efficient than I would like.

I'd also like to optimize the makeGuessList method, and make it easier on the eyes. I have tried to make it work recursively, but I have yet to succeed. I did create a nicer looking one, but it is less optimized and if the maxValue is low can take a lot of time to remove bad guesses. It creates two valeus as a lowEnd and highEnd for a list comprehension, then it creates a list of digital values not allowed, and finally runs through the whole list removing anything with those forbidden values.

Another area to optimize is the guessList itself. As the problem will accept the factors in any order, 223 is the same as 232 and 322. It seems a waste to have three entries when one will work fine. This would also reduce the work load needed when creating the rather intensive productList. I intend to write a part onto makeGuessList that goes through each entry, splits and sorts the string, then recombines the string, and then removes any duplicates. On second thought, I can possibly save a small amount of time by not recombining the string and leaving it as a sorted list. The for-loops used throughout will work on both.